[Refactor] Re-org index directories
