+++ be/src/common/config.h
@@ -1391,7 +1391,7 @@ CONF_mBool(enable_core_file_size_optimization, "true");
 // 6. query_rpc_thread_pool
 // use commas to separate:
 // * means release all above
-CONF_mString(try_release_resource_before_core_dump, "storage_page_cache")
+CONF_mString(try_release_resource_before_core_dump, "storage_page_cache");
 
 // Experimental feature, this configuration will be removed after testing is complete.
 CONF_mBool(lake_enable_alter_struct, "true");
