+++ be/src/storage/index/inverted/inverted_index_iterator.h
@@ -15,8 +15,8 @@
 #pragma once
 #include <string>
 
-#include "storage/inverted/inverted_index_option.h"
-#include "storage/inverted/inverted_reader.h"
+#include "storage/index/inverted/inverted_index_option.h"
+#include "storage/index/inverted/inverted_reader.h"
 
 namespace starrocks {
