+++ be/src/storage/index/inverted/inverted_index_option.cpp
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "storage/inverted/inverted_index_option.h"
+#include "storage/index/inverted/inverted_index_option.h"
 
 #include <boost/algorithm/string/case_conv.hpp>
