+++ be/src/storage/rowset/column_reader.cpp
@@ -47,8 +47,8 @@
 #include "common/logging.h"
 #include "runtime/types.h"
 #include "storage/column_predicate.h"
-#include "storage/inverted/index_descriptor.hpp"
-#include "storage/inverted/inverted_plugin_factory.h"
+#include "storage/index/index_descriptor.h"
+#include "storage/index/inverted/inverted_plugin_factory.h"
 #include "storage/rowset/array_column_iterator.h"
 #include "storage/rowset/binary_dict_page.h"
 #include "storage/rowset/bitmap_index_reader.h"
