+++ be/src/storage/rowset/column_writer.cpp
@@ -45,8 +45,8 @@
 #include "fs/fs.h"
 #include "gutil/strings/substitute.h"
 #include "simd/simd.h"
-#include "storage/inverted/inverted_index_option.h"
-#include "storage/inverted/inverted_plugin_factory.h"
+#include "storage/index/inverted/inverted_index_option.h"
+#include "storage/index/inverted/inverted_plugin_factory.h"
 #include "storage/rowset/array_column_writer.h"
 #include "storage/rowset/bitmap_index_writer.h"
 #include "storage/rowset/bitshuffle_page.h"
